resolving PR comments, moving all files to tests firectory
