+++ src/.env.example
@@ -1,2 +1,3 @@
+# This file show the ENV properties needed for local test execution
 DEV_CRI_CIC_API_URL=
-DEV_IPV_STUB_URL=
+DEV_IPV_STUB_URL=
